restructure simulations module into a package
